Add '.' character to project name regex
When creating dotnet core applications with the cli it allows
the character '.' and this is very common practice with
most dotnet applications.

The '.' is now allowed in the regex for project names when
creating. The behaviour is consistent with the cli by allowing
the user to add a '.' at the beginning or end of the project name.